spring security

+++ src/main/java/com/practicesecurity/messages/MessageService.java
@@ -4,22 +4,29 @@ package com.practicesecurity.messages;
 import com.practicesecurity.user.User;
 import com.practicesecurity.user.UserService;
 import org.springframework.stereotype.Service;
+import com.practicesecurity.security.CurrentUserService;
 
 import java.util.List;
 
 @Service
 public class MessageService {
 
+
     private final MessageRepo messageRepo;
     private final UserService userService;
 
-    public MessageService(MessageRepo messageRepo, UserService userService) {
+    private final CurrentUserService currentUserService;
+
+    public MessageService(MessageRepo messageRepo, UserService userService, CurrentUserService currentUserService) {
         this.messageRepo = messageRepo;
         this.userService = userService;
+        this.currentUserService = currentUserService;
     }
 
-    public List<MessageResponse> findByUser(String username) {
+    public List<MessageResponse> findByUser() {
 
+
+        String username = currentUserService.getUsername();
         User user = userService.getUserByUsername(username); // throws if not found
 
         return messageRepo.findByUser(user).stream()
@@ -38,20 +45,35 @@ public class MessageService {
         return messageRepo.findAll();
     }
 
-    public void createMessage(String username ,MessageRequest req) {
+    public Message createMessage(MessageRequest req) {
 
-        User user = userService.getUserByUsername(username);
+        String Username = currentUserService.getUsername();
+
+
+        User user = userService.getUserByUsername(Username);
+
+
         Message msg = new Message();
         msg.setMessage(req.getMessage());
         msg.setUser(user);
 
         messageRepo.save(msg);
 
+        return msg;
+
     }
 
     public void deleteMessage(Long id) {
 
-        Message msg = messageRepo.findById(id).orElseThrow(()->new RuntimeException( "Message Not Found"));
+        String Username = currentUserService.getUsername();
+
+        User currentUser = userService.getUserByUsername(Username);
+
+        Message msg = messageRepo.findById(id).orElseThrow(() -> new RuntimeException("Message not found"));
+
+        if(!msg.getUser().getId().equals(currentUser.getId())) {
+            throw new RuntimeException("Forbidden");
+        }
 
         messageRepo.delete(msg);
